# Speed up FERC 714 hourly demand transform